refactor/store-handling: Store hangling refactored/cleaned up.
