+++ app/extensions/safe/index.js
@@ -1,59 +1,19 @@
-import { session, shell } from 'electron';
 import logger from 'logger';
-import { CONFIG, isRunningProduction, SAFE } from 'appConstants';
+import { isRunningProduction, SAFE } from 'appConstants';
 import setupRoutes from './server-routes';
 import registerSafeProtocol from './protocols/safe';
-import {
-    saveConfigToSafe,
-    readConfigFromSafe
-} from './network/manageBrowserConfig';
+
 import registerSafeAuthProtocol from './protocols/safe-auth';
 import ipc from './ffi/ipc';
 
-import { initAnon, initMock, requestAuth, clearAppObj } from './network';
-import * as tabsActions from 'actions/tabs_actions';
-import * as safeActions from 'actions/safe_actions';
-import * as notificationActions from 'actions/notification_actions';
-import { urlIsAllowed } from './utils/safeHelpers';
+import { initAnon, initMock } from './network';
+// import * as tabsActions from 'actions/tabs_actions';
+
 import * as authAPI from './auth-api';
 
+import blockNonSAFERequests from './blockNonSafeReqs';
+import handleStoreChanges from './handleStoreChanges';
 
-const isForLocalServer = ( parsedUrlObject ) =>
-    parsedUrlObject.protocol === 'localhost:' || parsedUrlObject.hostname === '127.0.0.1';
-
-const blockNonSAFERequests = () =>
-{
-    const filter = {
-        urls : ['*://*']
-    };
-
-    const safeSession = session.fromPartition( CONFIG.SAFE_PARTITION );
-
-    safeSession.webRequest.onBeforeRequest( filter, ( details, callback ) =>
-    {
-        if ( urlIsAllowed( details.url ) )
-        {
-            logger.debug( `Allowing url ${details.url}` );
-            callback( {} );
-            return;
-        }
-
-        if ( details.url.indexOf( 'http' ) > -1 )
-        {
-            try
-            {
-                shell.openExternal( details.url );
-            }
-            catch ( e )
-            {
-                logger.error( e );
-            }
-        }
-
-        logger.info( 'Blocked req:', details.url );
-        callback( { cancel: true } );
-    } );
-};
 
 const init = async ( store ) =>
 {
@@ -89,10 +49,7 @@ const init = async ( store ) =>
 
     store.subscribe( async () =>
     {
-        manageSaveStateActions( store );
-        manageReadStateActions( store );
-        manageAuthorisationActions( store );
-        manageLogout( store );
+        handleStoreChanges( store );
     } );
 };
 
@@ -109,205 +66,6 @@ const init = async ( store ) =>
 //     // return next( action );
 // };
 //
-
-
-const authingStates = [
-    SAFE.APP_STATUS.TO_AUTH,
-    SAFE.APP_STATUS.AUTHORISING,
-    SAFE.APP_STATUS.AUTHORISATION_FAILED,
-    SAFE.APP_STATUS.AUTHORISATION_DENIED
-];
-
-
-const networkIsConnected = ( state ) =>
-{
-    const safeNetwork = state.safeNetwork;
-
-    if ( safeNetwork.appStatus === null || safeNetwork.networkStatus === SAFE.NETWORK_STATE.UNKNOWN ||
-        safeNetwork.networkStatus === SAFE.NETWORK_STATE.DISCONNECTED )
-    {
-        return false
-    }
-    else
-    {
-        return true;
-    }
-}
-
-/**
- * Handle triggering actions and related functionality for saving to SAFE netowrk
- * based upon the application stateToSave
- * @param  {Object} state Application state (from redux)
- */
-const manageReadStateActions = async ( store ) =>
-{
-    const state = store.getState();
-    const safeNetwork = state.safeNetwork;
-
-    if ( safeNetwork.readStatus !== SAFE.READ_STATUS.TO_READ )
-    {
-        // do nothing
-        return;
-    }
-
-    if ( safeNetwork.appStatus === SAFE.APP_STATUS.AUTHORISED &&
-        safeNetwork.networkStatus === SAFE.NETWORK_STATE.CONNECTED )
-    {
-        store.dispatch( safeActions.setReadConfigStatus( SAFE.READ_STATUS.READING ) );
-        readConfigFromSafe( store )
-            .then( savedState =>
-            {
-                store.dispatch( safeActions.receivedConfig( savedState ) );
-                store.dispatch(
-                    safeActions.setReadConfigStatus( SAFE.READ_STATUS.READ_SUCCESSFULLY )
-                );
-                return null;
-            } )
-            .catch( e =>
-            {
-                logger.error( e );
-                store.dispatch(
-                    safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.FAILED_TO_READ )
-                );
-                throw new Error( e );
-            } );
-    }
-    else if ( !networkIsConnected( state ) )
-    {
-        store.dispatch( safeActions.setReadConfigStatus( SAFE.READ_STATUS.FAILED_TO_READ ) );
-        store.dispatch( notificationActions.addNotification(
-            {
-                text: 'Cannot read. Network not yet connected.',
-                type: 'error'
-            } ) );
-
-    }
-    // TODO: Refactor and DRY this out between save/read?
-    else if ( !authingStates.includes( safeNetwork.appStatus ) )
-    {
-        store.dispatch( safeActions.setAuthAppStatus( SAFE.APP_STATUS.TO_AUTH ) );
-    }
-
-
-    // TODO: Refactor this: Is it needed?
-    if ( safeNetwork.readStatus === SAFE.READ_STATUS.FAILED_TO_READ &&
-        safeNetwork.appStatus === SAFE.APP_STATUS.AUTHORISED )
-    {
-        store.dispatch( safeActions.setReadConfigStatus( SAFE.READ_STATUS.TO_READ ) );
-    }
-};
-
-
-/**
- * Handle triggering actions and related functionality for Authorising on the SAFE netowrk
- * based upon the application auth state
- * @param  {Object} state Application state (from redux)
- */
-const manageAuthorisationActions = async ( store ) =>
-{
-    const state = store.getState();
-
-    if ( state.safeNetwork.appStatus === SAFE.APP_STATUS.TO_AUTH )
-    {
-        store.dispatch( safeActions.setAuthAppStatus( SAFE.APP_STATUS.AUTHORISING ) );
-        const app = await requestAuth();
-    }
-};
-
-/**
- * Handle triggering actions and related functionality for Authorising on the SAFE netowrk
- * based upon the application auth state
- * @param  {Object} state Application state (from redux)
- */
-const manageLogout = async ( store ) =>
-{
-    const state = store.getState();
-
-    if ( state.safeNetwork.appStatus === SAFE.APP_STATUS.TO_LOGOUT )
-    {
-        store.dispatch( safeActions.setAuthAppStatus( SAFE.APP_STATUS.LOGGING_OUT ) );
-        store.dispatch( safeActions.resetStore() );
-        clearAppObj();
-        store.dispatch( safeActions.setAuthAppStatus( SAFE.APP_STATUS.LOGGED_OUT ) );
-    }
-};
-
-
-/**
- * Handle triggering actions and related functionality for saving to SAFE netowrk
- * based upon the application stateToSave
- * @param  {Object} state Application state (from redux)
- */
-const manageSaveStateActions = async ( store ) =>
-{
-    const state = store.getState();
-    const safeNetwork = state.safeNetwork;
-
-    if ( safeNetwork.saveStatus !== SAFE.SAVE_STATUS.TO_SAVE )
-    {
-        // do nothing
-        return;
-    }
-
-    if ( networkIsConnected( state ) && safeNetwork.readStatus !== SAFE.READ_STATUS.READ_SUCCESSFULLY &&
-        safeNetwork.readStatus !== SAFE.READ_STATUS.READ_BUT_NONEXISTANT )
-    {
-        if ( safeNetwork.readStatus !== SAFE.READ_STATUS.TO_READ &&
-            safeNetwork.readStatus !== SAFE.READ_STATUS.READING )
-        {
-            logger.verbose( 'Can\'t save, not read yet. Triggering a read.' );
-            store.dispatch( safeActions.setReadConfigStatus( SAFE.READ_STATUS.TO_READ ) );
-        }
-
-        return;
-    }
-
-    if ( safeNetwork.appStatus === SAFE.APP_STATUS.AUTHORISED &&
-        safeNetwork.networkStatus === SAFE.NETWORK_STATE.CONNECTED )
-    {
-        store.dispatch( safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.SAVING ) );
-        saveConfigToSafe( store )
-            .then( () =>
-            {
-                store.dispatch(
-                    safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.SAVED_SUCCESSFULLY )
-                );
-
-                return null;
-            } )
-            .catch( e =>
-            {
-                logger.error( e );
-
-                // TODO: Handle errors across the store in a separate error watcher?
-                store.dispatch(
-                    safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.FAILED_TO_SAVE )
-                );
-                throw new Error( e );
-            } );
-    }
-    else if ( !networkIsConnected( state ) )
-    {
-        store.dispatch( safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.FAILED_TO_SAVE ) );
-        store.dispatch( notificationActions.addNotification(
-            {
-                text: 'Cannot save. Network not yet connected.',
-                type: 'error'
-            } ) );
-
-    }
-    else if ( !authingStates.includes( state.safeNetwork.appStatus ) )
-    {
-        store.dispatch( safeActions.setAuthAppStatus( SAFE.APP_STATUS.TO_AUTH ) );
-    }
-
-
-    if ( safeNetwork.saveStatus === SAFE.SAVE_STATUS.FAILED_TO_SAVE &&
-        safeNetwork.appStatus === SAFE.APP_STATUS.AUTHORISED )
-    {
-        store.dispatch( safeActions.setSaveConfigStatus( SAFE.SAVE_STATUS.TO_SAVE ) );
-    }
-};
 
 
 export default {
